% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ces-package.R
\docType{package}
\name{ceshat-package}
\alias{ceshat-package}
\alias{_PACKAGE}
\alias{ceshat}
\title{ceshat: Nonparametric estmation of Conditional ES}
\description{
ceshat provides estimation functions for CVaR and CES.
}
\author{
Young-geun Kim <\url{https://github.com/ygeunkim}>
}
\keyword{internal}
