% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.R
\name{compute_gh}
\alias{compute_gh}
\title{Distribution function of kernel}
\usage{
compute_gh(x, type = c("Gaussian", "Epanechinikov", "Tricube", "Boxcar"), h)
}
\arguments{
\item{x}{point where the kernel function is evaluated}

\item{type}{which kernel to use? \code{c("Gaussian", "Epanechinikov", "Tricube", "Boxcar")}.}

\item{h}{bandwidth}
}
\value{
Distribution function value
}
\description{
Distribution fuction of \code{\link{compute_kernel}}
}
