% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.R
\name{compute_kernel}
\alias{compute_kernel}
\title{Symmetric kernel function}
\usage{
compute_kernel(
  x,
  type = c("Gaussian", "Epanechinikov", "Tricube", "Boxcar"),
  h
)
}
\arguments{
\item{x}{point where the kernel function is evaluated}

\item{type}{which kernel to use? \code{c("Gaussian", "Epanechinikov", "Tricube", "Boxcar")}.}

\item{h}{bandwidth}
}
\value{
Gaussian function value given \code{x}
}
\description{
Kernel function with bandwidth
}
\details{
Gaussian kernel function.
\deqn{K_h(x) = \frac{1}{\sqrt{2\pi h}} \exp( - \frac{x^2}{2h})}
}
