% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytic.R
\name{invert_cvar}
\alias{invert_cvar}
\title{Inverting 1 - conditional CDF}
\usage{
invert_cvar(p = 0.95, cdf, x, lower = -1, upper = 1)
}
\arguments{
\item{p}{Upper tail probability. \code{0.95}, by default.}

\item{cdf}{conditional cdf given x. Function of \code{(y, x)}}

\item{x}{conditional information variable \code{x}}

\item{lower}{lower value to be primarily searched}

\item{upper}{upper value to be primarily searched}
}
\value{
CVaR value for each \code{x}
}
\description{
This function computes true CVaR given true conditional CDF.
}
\details{
Note that
\deqn{\hat{nu}_p(x) = \hat{S}_c^{-1}(p \mid x)}
where
\deqn{\hat{S}(y \mid x)_c(y \mid x) = 1 - \hat{F}_c(y \mid x)}
Inverting can be done by finding
\deqn{\inf \{ y : F(y \mid x) \ge 1 - p \}}
When finding the minimum, firstly find in \code{(lower, upper)}.
But the answer can be out of this bound.
These arguments just set the primary interval, not the final one.
}
\references{
Cai, Z., & Wang, X. (2008). \emph{Nonparametric estimation of conditional VaR and expected shortfall}. Journal of Econometrics, 147(1), 120-130.
}
