% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.R
\name{kernel_gauss}
\alias{kernel_gauss}
\title{Gaussian kernel function}
\usage{
kernel_gauss(x, h)
}
\arguments{
\item{x}{point where the kernel function is evaluated}

\item{h}{bandwidth}
}
\value{
Gaussian function value given \code{x}
}
\description{
Gaussian kernel function with bandwidth
}
\details{
\code{h} is the standard deviation.
\deqn{K_h(x) = \frac{1}{\sqrt{2\pi h}} \exp( - \frac{x^2}{2h})}
}
