% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytic.R
\name{plugin_ces}
\alias{plugin_ces}
\title{Plug-in Method}
\usage{
plugin_ces(p = 0.95, pdf, cdf, x, lower = -2, upper = 2)
}
\arguments{
\item{p}{Upper tail probability. \code{0.95}, by default.}

\item{pdf}{conditional pdf given x. Function of \code{(y, x)}}

\item{cdf}{conditional cdf or cvar}

\item{x}{conditional information variable \code{x}}

\item{lower}{lower value to be searched}

\item{upper}{upper value to be searched}
}
\value{
CES function of \code{x}
}
\description{
This function computes true CES given true CVaR and conditional pdf.
}
\details{
Note that
\deqn{\hat{\mu}_p(x) = \frac{1}{p} \sum_{t = 1}^n W_{c,t}(x, h) \left[ Y_t \bar{G}_{h_0} (\hat{\nu}_p (x) - Y_t) + h_0 G_{1, h_0} (\hat{\nu}_p (x) - Y_t) \right]}
}
\references{
Cai, Z., & Wang, X. (2008). \emph{Nonparametric estimation of conditional VaR and expected shortfall}. Journal of Econometrics, 147(1), 120-130.
}
