% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_cai.R
\name{predict.ces}
\alias{predict.ces}
\title{Predict method for WDKLL ces}
\usage{
\method{predict}{ces}(object, newx, ...)
}
\arguments{
\item{object}{Object of class from \code{\link{wdkll_ces}}}

\item{newx}{x to predict. Unless specified, use the \code{data}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
WDKLL values for CES
}
\details{
Plugging-in in methods gives
\deqn{\hat{\mu}_p(x) = \frac{1}{p} \sum_{t = 1}^n W_{c,t}(x, h) \left[ Y_t \bar{G}_{h_0} (\hat{\nu}_p (x) - Y_t) + h_0 G_{1, h_0} (\hat{\nu}_p (x) - Y_t) \right]}
}
\references{
Cai, Z., & Wang, X. (2008). \emph{Nonparametric estimation of conditional VaR and expected shortfall}. Journal of Econometrics, 147(1), 120-130.
}
