% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wnw.R
\name{predict.nwcvar}
\alias{predict.nwcvar}
\title{Predict method for nwcvar}
\usage{
\method{predict}{nwcvar}(object, newx, nw = FALSE, ...)
}
\arguments{
\item{object}{Object of class from \code{\link{wdkll_cvar}}}

\item{newx}{x to predict. Unless specified, use the \code{data}.}

\item{nw}{Use NW or WNW. \code{TRUE} if NW.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
CVaR given \code{x}
}
\description{
WDKLL values for CVar
}
\details{
CVaR can be earned by inverting the CDF.
\deqn{\hat{nu}_p(x) = \hat{S}_c^{-1}(p \mid x)}
where
\deqn{\hat{S}(y \mid x)_c(y \mid x) = 1 - \hat{F}_c(y \mid x)}
}
\references{
Cai, Z., & Wang, X. (2008). \emph{Nonparametric estimation of conditional VaR and expected shortfall}. Journal of Econometrics, 147(1), 120-130.
}
