% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wnw.R
\name{wnw_cvar}
\alias{wnw_cvar}
\title{Weighted Double Kernel Local Linear Estimation of Conditional Value at Risk}
\usage{
wnw_cvar(
  formula,
  data,
  prob = 0.95,
  nw_kernel = c("Gaussian", "Epanechinikov", "Tricube", "Boxcar"),
  nw_h,
  init = 0,
  eps = 1e-05,
  iter = 1000,
  lower_invert = -3,
  upper_invert = 3
)
}
\arguments{
\item{formula}{an object class \link[stats]{formula}.}

\item{data}{an optional data to be used.}

\item{prob}{upper tail probability for VaR}

\item{nw_kernel}{Kernel for weighted nadaraya watson}

\item{nw_h}{Bandwidth for WNW}

\item{init}{initial value for finding lambda}

\item{eps}{small value}

\item{iter}{maximum iteration when finding lambda}

\item{lower_invert}{lower y when inverting the cdf}

\item{upper_invert}{upper y when inverting the cdf}
}
\value{
CVaR given \code{x}
}
\description{
WDKLL estimator of CVaR
}
\details{
CVaR can be earned by inverting the CDF.
\deqn{\hat{nu}_p(x) = \hat{S}_c^{-1}(p \mid x)}
where
\deqn{\hat{S}(y \mid x)_c(y \mid x) = 1 - \hat{F}_c(y \mid x)}
}
\references{
Cai, Z., & Wang, X. (2008). \emph{Nonparametric estimation of conditional VaR and expected shortfall}. Journal of Econometrics, 147(1), 120-130.
}
