% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceshat-package.R
\docType{package}
\name{ceshat-package}
\alias{ceshat-package}
\alias{_PACKAGE}
\alias{ceshat}
\title{ceshat: Nonparametric Estimation of Conditional Expected Shortfall}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Nonparametric ways to estimate conditional value at risk (expected shortfall),
    and conditional expected shortfall.
}
\details{
ceshat provides estimation functions for CVaR and CES.
}
\references{
Cai, Z., & Wang, X. (2008). \emph{Nonparametric estimation of conditional VaR and expected shortfall}. Journal of Econometrics, 147(1), 120-130.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ygeunkim}
}

}
\keyword{internal}
