% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cai_ces.R
\name{wdkll_cdf}
\alias{wdkll_cdf}
\title{Weighted Double Kernel Local Linear Estimation of Coditional CDF}
\usage{
wdkll_cdf(
  formula,
  data,
  nw_kernel = c("Gaussian", "Epanechinikov", "Tricube", "Boxcar"),
  nw_h,
  pdf_kernel = c("Gaussian", "Epanechinikov", "Tricube", "Boxcar"),
  h0,
  init = 0,
  eps = 1e-05,
  iter = 1000
)
}
\arguments{
\item{formula}{an object class \link[stats]{formula}.}

\item{data}{an optional data to be used.}

\item{nw_kernel}{Kernel for weighted nadaraya watson}

\item{nw_h}{Bandwidth for WNW}

\item{pdf_kernel}{Kernel for initial estimate of conditinal pdf}

\item{h0}{Bandwidth for pdf kernel}

\item{init}{initial value for finding lambda}

\item{eps}{small value}

\item{iter}{maximum iteration when finding lambda}
}
\value{
Conditional CDF function with argument \code{y} and \code{x}
}
\description{
This function estimates conditional CDF
using WDKLL method.
}
\details{
\deqn{\hat{F}_c(y \mid x) = \sum_{t = 1}^n W_{c,t} G_{h_0} (y - Y_t)}
}
\examples{
X <- data.frame(x = 1:10, y = rnorm(10))
wdkll_cdf(y ~ x, X, nw_h = 1, h0 = 1)(1, 1)
wdkll_cdf(y ~ x, X, nw_h = 1, h0 = 1)(2, 1)
}
\references{
Cai, Z., & Wang, X. (2008). \emph{Nonparametric estimation of conditional VaR and expected shortfall}. Journal of Econometrics, 147(1), 120-130.
}
